// frontend/src/app/generated-sources/openapi/model/seatingPlanStaticElement.ts
/**
 * Ticketline API
 * Documentation for the Ticketline API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ElementLocation } from './elementLocation';


export interface SeatingPlanStaticElement { 
    id: number;
    color: number;
    description?: string;
    location: ElementLocation;
}